requirements.txt
tensorflow>=2.5.0
numpy>=1.19.5
matplotlib>=3.4.2
pillow>=8.2.0
scikit-learn>=0.24.2
seaborn>=0.11.1
opencv-python>=4.5.2 